+++ .flake8
@@ -1,0 +1,11 @@
+[flake8]
+max-line-length = 88
+extend-ignore = E203, E501
+exclude = 
+    .git,
+    __pycache__,
+    build,
+    dist,
+    *.egg-info
+
+max-complexity = 10
