--- .github/workflows/test.yml
@@ -1,27 +1,0 @@
-name: PR Tests
-
-on:
-  pull_request:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-
-      - name: Install dependencies
-        run: |
-          pip install -r requirements.txt
-          pip install pytest
-
-      - name: Run tests
-        env:
-          PYTHONPATH: ${{ env.PYTHONPATH }}
-        run: |
-          pytest --verbose
