+++ Makefile
@@ -1,0 +1,8 @@
+launch:
+	@bash scripts/launch_chromium.sh
+
+stop:
+	@bash scripts/stop_chromium.sh
+
+restart:
+	@bash scripts/restart_chromium.sh
