+++ pyproject.toml
@@ -1,0 +1,29 @@
+[tool.poetry]
+name = "Job scrappers"
+version = "1.4.4"
+description = "A package to scrap job offers, send SMS notifications through a free SMS API, and transfer data to a Notion DB."
+authors = ["Marc <128506536+MarcChen@users.noreply.github.com>"]
+packages = [
+    { include = "src" }
+]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+selenium = "4.9.1"
+beautifulsoup4 = "4.12.2"
+lxml = "4.9.3"
+requests = "2.31.0"
+rich = "13.5.2"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.0.1"
+flake8 = "^7.1.1"
+black = "^22.3.0"
+isort = "^6.0.0"
+
+[tool.isort]
+profile = "black"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
