--- requirements.txt
@@ -1,5 +1,0 @@
-selenium==4.9.1
-beautifulsoup4==4.12.2
-lxml==4.9.3
-requests==2.31.0
-rich==13.5.2
