+++ scripts/stop_chromium.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+stop_chromium() {
+    docker stop selenium-chromium
+    docker rm selenium-chromium
+    echo "Chromium stopped"
+}
+
+stop_chromium
